requirements.txt
Flask==2.3.3
google-cloud-storage
